pyproject.toml
[project]
name = "content-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "langgraph>=1.1.3",
    "openai>=2.0.0",
    "ipykernel>=7.2.0",
    "llama_cloud>=1.0",
    "langchain-openai>=1.1.12",
]
